vmbus_server: save reserved channels when disconnected. (#896)
The vmbus server only saved channel state when connected. Since reserved
channels can be open even if the guest is not currently connected to
host vmbus, this creates a small window where reserved channels can be
lost.

This change separately saves reserved channels when the guest is
disconnected. This is done only when disconnected; when connected,
reserved channels are saved with the connected state as before, to
maintain compatibility with the saved state of versions prior to this
fix.

Fixes #882.