flowey: use baseline artifact instead of igvm-extras for binary comparison (#906)
As part of #475, I added a new baseline artifact for binary comparison
but didn't use it yet as I wanted to wait for there to be builds to diff
against but hadn't followed up yet. #895 renamed the binary in
igvm-extras that we were using to diff and broke the comparison.

This change uses the baseline artifact for comparison and keeps the
binary size artifacts limited in scope.