openhcl: refactor all dma users to use dma_manager (#898)
Previously, there were multiple ways to allocate pages intended for DMA
throughout OpenHCL, with policy spread throughout the code. This change
refactors all devices and other users such as the partition to construct
clients via the `dma_manager` inside `LoadedVm`, with the specified
policy for each client provided at creation time. The new
`openhcl_dma_manager` crate contains the logic on how that client policy
is mapped to the various sources of memory available for DMA
allocations.

As part of this change, the shared and private page pools ownership are
moved to `dma_manager`, with a simple save restore implementation to
support the necessary save restore for nvme keep alive.

This should help future changes be limited in scope to just the
`openhcl_dma_manager` crate, such as upcoming pin/unpin support for dma
transactions.

Additionally, this changes the behavior of the AK cert renewal for the
GET to not require the private pool, as since persistence is not needed,
along with contiguous page allocations not required.

Closes #786