xtask: fix interplay between fmt-on-save and xsync (#925)
Fixing the underlying issue from #914

This PR does a few things:

1. The `xtask-path` logic has been fixed to always drop the file under
`target/xtask-path`, regardless of whether a `--custom-root` is being
used or not.
2. A new `--run-on-save` parameter has been added to `xtask`, which can
be used to skip certain expensive checks when running `xtask` in
response to a file save event. It is not currently being used by the
open-source `xtask`, but plays an important role in the internal repo's
`xtask fmt` implementation.
3. Updates the Guide to make it more clear why one may or may not want
to opt into this functionality.

**These changes are not particularly interesting in isolation, but make
more sense when reviewed alongside the changes being made in the
internal repo.**

The fact we have this sort of coupling between the oss xtask and the
closed-source xtask is unfortunate, and something we should try and move
away from in the future...