vmbus_server: support hvsocket with vmbusproxy. (#959)
This change updates vmbusproxy to correctly handle hvsocket channel
offers. It also allows it to pass connection requests from the guest
(the `TlConnectRequest` message) to the proxy driver, and for the driver
to request a response (`TlConnectRequestResult`) be sent to the guest.

This functionality is not enabled for openvmm, as it is mutually
exclusive with openvmm's normal hvsocket handling, and requires
vmbusproxy.sys driver updates.