netvsp: Move all worker start logic to coordinator to avoid races (#958)
If a netvsp worker task is stopped waiting for external processing,
avoid restarting it due to other signals. This is mainly a problem for
the primary worker task which will stop itself to wait for work to be
done. If some other trigger starts it before that work is complete, then
it is in an invalid state. Move all worker start logic to a single
location in the coordinator.

For the primary worker make sure there are no outstanding messages to
process before starting it, as most messages will stop it again anyway,
but more importantly one of the outstanding messages may be from the
primary worker, requesting work to be done before starting it up again.

---------

Co-authored-by: Brian Perkins <brian.perkins@microsoft.com>