vmbus_client: make Close an RPC (#972)
The vmbus client and vmbus relay channel states need to stay in sync,
since we cannot save/restore in intermediate states (such as in the
middle of opening a channel).

Close currently breaks this--the relay can send a close just before
pausing saving, and there is no guarantee (that I can see) that the
client will process it before it is saved.

Fix this by making `Close` an RPC, so that the relay can wait for the
message to be sent to the host, and the client state updated, before it
pauses and saves its own state.