vmbus_relay: split vmbus client into caller (#969)
Today, the vmbus relay owns the vmbus client. This makes it difficult to
use the vmbus client outside of the relay, necessitating the awkward
"intercept channel" feature. This works OK for the shutdown IC, but it
doesn't work well in scenarios without the relay or before the relay has
started.

To enable more flexible configurations, split the vmbus client out from
the relay. Store the client saved state as part of servicing state. To
support saved state backward and forward compatibility, also store the
client saved state in the relay's protobuf blob.