virt_mshv_vtl: Refactor TLB flush and lock code to allow for partial borrows (#943)
This allows us to remove the annoying 'delay-flush' struct I added as a
workaround previously. This is needed so we can perform similar types of
activities elsewhere without having to use the same delayed workaround,
like in https://github.com/microsoft/openvmm/pull/921.

This did require the removal of `vtls_tlb_waiting` in order to make all
the partial borrowing possible, however this field was only being used
as a small performance optimization to avoid a relaxed atomic load.
Since this load is relaxed, I think this is ok?