Use async closures everywhere (#930)
Now that Rust 1.85 is out, switch to using async closures instead of
closures returning async blocks. Use `AsyncFn*` instead of `Fn* -> impl
Future`.

In some places, this requires a little more verbosity: the brief `||
self.foo()` becomes `async || self.foo().await`. But in return, all the
weird lifetime issues go away, so we're able to cut down on `let x = &x`
and `async move`s that shouldn't be needed. In the case of the chipset
builder, this is a significant improvement--we are able to reference
various motherboard services within the async closure for adding a
device instead of having to pre-clone them outside the closure.

There are still a few places where we have to use the `FnOnce() -> Fut`
pattern since `AsyncFnOnce` doesn't yet have a mechanism for specifying
that the resulting future must be `Send`.