vmbus: add pause/resume messages, use them in vmbus_client save/restore (#962)
During VTL2 servicing, the host resets the VTL2 synic state, causing any
pending vmbus messages to be lost. To mitigate this, `vmbus_client`
masks the synic sint and drains and processes any pending messages
before saving its state. This prevents the vmbus server in the Hyper-V
host from posting new messages, so that they are not lost; Hyper-V
retries them later when the sint is unmasked.

Unfortunately, this mask/unmask behavior is not supported by OpenVMM
with `virt_whp`, because `virt_whp`'s synic emulation relies on
`HvPostMessageDirect`, which does not allow posting a message to the
message slot when the sint is masked. This means that it is impossible
for `virt_whp` to support this process for draining the message queue,
and so any pending messages are lost across VTL2 servicing.

Perhaps we should fix the hypervisor to support posting a message with
`HvPostMessageDirect` even when the sint is masked, and to provide a
mechanism to get an exit in `virt_whp` when the sint becomes masked so
that we can stop queuing new messages, to match the behavior of Hyper-V
precisely.

However, it is arguably cleaner just to support pausing/resuming the
vmbus message stream in-band in the vmbus protocol. This works around
this issue and it avoids the need for the host to get a "sint unmasked"
notification or to otherwise retry on a timer.

So, implement that. Add pause, pause complete, and resume messages to
the vmbus protocol, behind a protocol feature flag, to allow the client
to tell the server to temporarily stop sending synic messages. Use this
feature when available in `vmbus_client`; when not available, fall back
to the synic masking approach.

Fixes #954. Reverts #955.