virt_mshv_vtl: Cleanup some vpindexes (#980)
Try to use VpIndex everywhere instead of usize and u32. The main place
where we aren't today is in the hypercall traits, which are defined
elsewhere.

This also removes a TODO about cpu_index potentially being different
than vp_index, and the cpu_index field entirely, under the assumption
that if it hasn't been an issue up till now it's not going to be. If we
do want to keep this field we should probably have some docs and a more
careful audit of when we should be using vp_index vs cpu_index.