vmbus_client: fix state machine around revoke (#977)
When the server revokes a channel:

* Don't allow the channel ID to be reused until the user of the channel
releases it. This ensures the server cannot overwhelm the client and its
users with channel offers and that we won't see any duplicate offers in
either the client or the downstream relay/server.

* Ensure channel requests can continue to be issued without
unpredictable behavior.

* Save a channel ID release message for the channel if the client is
saved before the user releases the revoked channel.

Add some tests to cover these cases.