Enlighten the Windows cross tool about midlrt (#998)
This change adds `midlrt.exe` to the tools supported by
`cross_tool.py`. This is a Windows tool included with the VS tooling and
is executed in Windows. As such, when executing this tool from within
WSL, we set `WSLENV` to pass in the required `LIB` and `INCLUDE` env
vars. Also, add an `sdk` value to the cross tool dump which prints the
location of the Windows SDK.

Also updated is `setup_windows_cross.sh` and the windows cross `config.toml` to add environment variables
for the location of midlrt and the cross tool itself, which will be
needed by some Rust build scripts.