hvdef: clean up `HvMessage` type and handling (#1036)
Align `HvMessage` to 16 bytes since it sometimes contains messages from
the hypervisor that have 16-byte alignment, and it is always 16-byte
aligned in mapped memory interfaces.

Add new `as_message` method for casting the payload to a message type
without having to use an explicit zerocopy dance. This method validates
that the type being cast to is supposed to be a message type and that it
has the proper size and alignment.

Use `as_message` in virt_mshv_vtl.