virt_mshv_vtl: Bring back BackingPrivate (#1046)
As requested. This can't be quite as clean as hcl's BackingPrivate due
to leaking out the associated StateAccess type through our impl of
virt::Processor requiring BackingPrivate to be fully pub. I played
around with the types for a while but couldn't get anything that was
better than this. So just leave the expects in.