virt_mshv_vtl: use BTreeMap for cvm cpuid subleafs (#1051)
XSAVE size and alignment calculation requires enumerating subleafs in
ascending order. Previously, using a hashmap means that iteration order
is not stable, so switch it to a BTreeMap.

A follow up change will remove this path entirely, but take this small
fix to fix main. This is the fix being taken back to release/2411.

#708