For tx segments exceeding 31, coalesce into bounce buffer (#968)
net_mana has a hardware limit of on the number of segments (either 31 or
30).

Any packet with fewer than 31 segments, net_mana will send correctly.
When a packet has a number of segments greater than 31, openhcl crashes.

Note: netvsp has a max MDL chain length of 34, so for packets with an
MDL chain longer than 34, netvsc will bounce buffer, and net_mana will
see only 2 segments. This makes the crash rare / difficult to reproduce.

Fixing the problem by:
* Modifying net_mana code to bounce buffer whenever segment count
exceeds the hardware limit.
* Adding functionality to ContiguousBufferManager to create logic for
transactions and try_extend to support a single bounce buffer that can
include either:
    * Header's bounced segments + beginning segments of payload.
    * The first sequence of the payload that can be coalesced.
* Added a new unit test to net_mana to validate coalescing by sending 34
segments.

There is a limitation on the size of the bounce buffer of 4096. It is
possible that coalescing will not succeed at getting the number of
segments below the hardware limitation. In those cases, the Packet is
dropped and an error is traced "Failed to bounce buffer the packet too
many segments".

Test results:
112,200 packets with different combinations of # MDLs and various
payload sizes sent.
0 crashes
 
680 (.6%) error "Failed to bounce buffer the packet too many segments"
segments_remaining=0x1 350-ish
segments_remaining=0x2 321-ish
*these numbers are approximate because the error message were rate
limited and some were dropped, but it's good enough for napkin math.