user_driver: moving emulated.rs file to its own crate and updating EmulatedDevice to accept a generic allocator (#945)
The `emulated.rs` file in the `user_driver` crate provides some wrapper
structs around memory and user_driver devices and in turn provides
functionality for testing the devices such as `NvmeController`. This
file used to be in `user_driver` but needed to be moved to its own crate
(`user_driver_emulated_mock`) since it is providing testing-only
functionality. Going forward, this will allow crates to use the
`emulated.rs` functions/structs as [dev-dependencies].
The new `user_driver_emulated_mock` crate futher breaks up the contents
in `emulated.rs` into separate files for better organization of structs
implementing unsafe code.
The `EmulatedDevice` type that is being leveraged for driver (nvme,
mana) unit tests in OpenVMM currently uses the `EmulatedDmaAllocator`
behind the scenes. This is causing issues during `save_restore` testing
since the `attach_dma_buffer` fn is not implemented for
`EmulatedDmaAllocator`.
This change will allow the `EmulatedDevice` to accept and use the
`PagePool` allocator instead as it can now accept any type that
implements the `DmaClient` trait.