linting: Turn on check-private-items, clean up safety docs (#1070)
This PR turns on clippy's check-private-items option, which causes some
documentation lints to also check private items, not just public. For
unsafe code this is desired, as even if the functions are not publicly
exposed, we should still document the safety requirements for our own
sake, as we change these libraries all the time. This PR adds those
comments where it was possible for me to do so and adds a few new
exclusions where it wasn't.

Cleaning up the remaining missing safety docs on crates that ship as
part of prod is tracked by
https://github.com/microsoft/openvmm/issues/1069. Cleaning up the
remaining missing safety docs on crates that do not ship as part of prod
is not currently tracked. We should ensure as we bring up new platforms
(eg kvm, whp, mshv) that we clean up their documentation while
productizing them.