repo: Clean up some more allows (#1076)
The vast majority of the remaining allows are related to naming
conventions, which seem to be bugged at the moment (expect will claim
they're unfulfilled but removing it will have them appear). Then there's
also a few that would require complicated preconditions, or are
macro-generated, which likely will remain forever. Though possibly some
more of the macro-generated ones can go away, we'll see.

This is cargo-hack clean.