user_driver_emulated_mock: remove `DeviceSharedMemory` and related structs. Instead use `DeviceTestMemory` backed by `TestMapper` (#1084)
A follow up PR to #945

This PR completely removes the `DeviceSharedMemory` (which is backed by
`AlignedHeapMemory`) and some of the related structs (such as
`EmulatedDmaBuffer`). It also transitions the tests using
DeviceSharedMemory over to using the newly created `DeviceTestMemory`
struct (which is backed by `TestMapper`). This only impacts testing for
Mana and Nvme drivers, both of which need to use `TestMapper` mapped
memory in order to create keepalive tests in the future.

As part of this transition, the preexisting tests for Nvme have also
been slightly simplified with added comments to explain the test wiring.