virt_mshv_vtl: Optimize and cleanup cvm_handle_cross_vtl_interrupts (#1097)
Checking for VINA in every exit is hurting TDX perf due to the read of
VMX_VMCS_GUEST_INTERRUPTIBILITY. Tweak cvm_handle_cross_vtl_interrupts
so that we don't do anything if VTL 1 isn't yet enabled to avoid this as
much as we can. Also reorder the ifs so that this read is always checked
last. Also, thanks to how we've been moving around state everywhere,
flatten out these ifs so things aren't so deeply nested.