livedump: Don't include two copies of kmsg (#1106)
Both underhill-dump and underhill-crash will include a copy of kmsg in
the final dump, as they weren't originally intended to be used together.
Using them together works fine, but it means we have two copies of kmsg
in the output. This doesn't cause any problems with reading the dump, it
just wastes space. underhill-dump has the easier job here, so fix this
by adding yet another option to underhill-crash.