Fix lints for Rust 1.86 (#1131)
Mostly a bunch of manual div_ceils and a new doccomment formatting lint.
The one potentially controversial change I think is my decision to
remove needless_continue from our list of manually enabled
allow-by-default lints. This new version was allowing it to trigger in a
lot of new locations, locations that in my opinion benefit from the
local clarity the continue provides, even if it is strictly speaking not
needed. I'm open to keeping it around if anyone else feels strongly
about it.