hv1_emulator: Pass GuestMemory on construction instead of on access, allow a configurable number of VTLs (#1114)
This makes the structs themselves a little bigger, but makes interacting
with them cleaner IMO.

Making this work on WHP required moving the hvstate to the top-level
partition, instead of the per-vtl partition. Doing this required making
the emulator support a configurable number of VTLs, so that one single
top-level Hv can emulate for both VTL 0 and VTL 2.