vmbus_server: improve MnF support (#1073)
This change updates vmbus_server's support for monitored interrupts.

- The `SynicPortAccess` trait is updated to provide information about
monitor usage during host and guest event port creation. When creating
guest event ports, initial parameters for their creation must be
supplied.
- The `MonitorPortAccess` trait is updated to require the VTL that a
monitor page is for.
- The VmBus server delays creation of guest event ports until the
channel is opened, at which point all the information needed to create
it with initial parameters is known.
- Monitored interrupt latency is passed along for vmbusproxy channels,
and provided by internal channels using MnF (this is not used by
openvmm's implementation of MnF, since it is emulated, but can be used
by other implementations).