hv1_emulator: Lock overlay pages (#1109)
By locking the pages we are configured for we can guarantee there will
be no errors reading or writing to the page. This makes the code shorter
and easier to read.

If we like this approach I'll keep going with the rest of hv1_emulator,
just wanted to get this out for initial impressions.