hyperv_ic: add timesync implementation (#1173)
Add an implementation of the timesync integration service, used to
provide the wall clock time to the guest.

As part of this, expose the partition reference time consistently from
all the different hypervisor backends. The reference time is used in the
timesync IC message to provide a reference point for the guest to know
when the wall clock snapshot was taken.