vmbus_client: fix handling of revoked channels with active requests (#1100)
If there are active GPADL or modify channel requests when the host
revokes a channel, it will still send responses for those requests. This
is different from channel open, for which the host will only send a
response if the channel has not yet been revoked.

`vmbus_client` was not handling this properly, so it would release the
channel ID after receiving a revoke and before the GPADL responses
arrived. This would cause panics when the request responses eventually
arrived.

Fix this by tracking the state more carefully and only releasing the
channel once all active requests have been completed by the host. Wait
for such requests during pause so that we do not have to save channels
in a revoked-with-requests-outstanding state.

Also, fix a theoretical deadlock in the pause path. And now that #963
has been merged into the release branch, remove the workaround for older
releases not restoring pending messages.