flowey: add type-safe, automatic artifact publish/resolution (#1183)
Artifacts within flowey are used to share files between pipeline jobs,
and to publish build results for diagnostics and human consumption.
Today, the process of defining, publishing, and consuming artifacts is
manual--flowey gives you a path, and it's your responsibility to use it
correctly. We have some conventions that make this easier to get right
(manually define flowey nodes for publishing and resolving artifacts),
but they result in a lot of boilerplate code, which makes refactoring
tedious.

Add a new typed artifact concept, using `serde` to define the layout of
the artifact in the file system. Update most of the code to use this new
concept, leaving behind the artifacts that are "weird" in some way--I'll
clean those up in a subsequent change.