openhcl: add cmdline option to disable uefi frontpage (#1184)
Add the ability to do this via an environment variable. While the device
platform settings also specify this setting, not all hosts (such as
Hyper-V) support configuring this setting in all scenarios.

This is to be used in an upcoming petri change to match the correct
expected behavior for petri vm configuration.