underhill_core: add OPENHCL_SIGNAL_VTL0_STARTED option (#1213)
Add a new option, `OPENHCL_SIGNAL_VTL0_STARTED`, for signaling the host
that VTL0 is ready very early in paravisor start. This is useful when
you don't want a Hyper-V host to wait for VTL0 to start (e.g., because
it will never start or will delayed indefinitely via
`OPENHCL_WAIT_FOR_START`).