openhcl_boot: report partition memory with different device type to workaround kernel numa parsing (#1219)
The Linux kernel assumes any dt node that has a device-type of `memory`
is a memory node, even if it's not under the expected system memory path
of `/memory/...`. This causes Linux to attempt to add these node's numa
information to internal data structures, resulting in it failing to
parse any of the numa topology information for ranges that are actual
VTL2 memory for the kernel. We report under `/openhcl/memory@...` the
whole partition memory information for use by usermode.

Workaround this by reporting the partition memory information with a
different device type, `memory-openhcl`. Add a new test to make sure the
Linux kernel didn't run into any errors parsing numa information.