flowey: use KnownTestArtifacts instead of KnownVhds and KnownIsos (#1221)
Since flowey does not care whether an artifact is a VHD, ISO, or
something else, we can make this more loosely typed and simplify the
code. This also allows us to add different types of artifacts that have
different characteristics (like VMGS). The type information is still
enforced in the `petri_artifacts_vmm_test` crate.