virt_mshv_vtl/hv1_emulator: Move hypervisor cpuid construction into cvm_cpuid (#1169)
As previously discussed, this moves the creation of various CPUID leaves
into per-backing code inside virt_mshv_vtl, instead of being lumped
together in hv1_emulator. It also adds the previously missing
register_pat_available bit to both SNP and TDX, since we support reading
PAT. It also removes the bit for access_vp_runtime_msr, as it is not an
msr we implement anywhere.