TDX: Add support for descriptor table intercepts & some emulation (#1217)
Adds support for intercepting descriptor table access instructions and
generating intercept messages when appropriate. Because VMX only gives
us an 'all-or-nothing' bit to enable exits we have to emulate
instructions that generate an exit but were not actually asked for by
VTL 1. Adds support for half of these, those that touch GDT or IDT, with
the other half to come in a follow up because I need a brain break.
Their design will likely be very similar anyways.

Slightly refactor the instruction emulator to allow external emulation
of memory accesses, without going through the full decoding pipeline
first. This primitive could be useful if we end up having to emulate
more instructions like this for specific backends with decoding
assistance in the future.

Note that only the two IDT instructions seem to actually be run in my
current test workloads, so the GDT instructions are technically
untested.

With this PR a single VP TDX Guest VSM VM is able to fully boot, enable
credential guard, and be RDP'd into!