vmm_tests: enable TMK tests on Hyper-V (#1228)
Add a bunch of infra to petri's Hyper-V support so that we can run the
TMK tests.

Also add enough infra to run the sidecar tests, but don't turn those
tests on yet--they are failing. I'll debug and fix this separately.

Also, disable aarch64 for now, since the TMK VMM seems to hang.