loader: use the x86 selector definition (#1211)
The existing code works by luck because we use the ring 0
GDT selectors. The selector is computed by multiplying the
index by the size of the small GDT entry (by 8 bytes), and
the ring 0 GDT selectors can be computed by shifting by 3.
That gives the correct result by chance rather than by
following the specification.

Extend the x86 definitions, calculate the selectors as the
specification defines.