Revert "Build musl binaries with RELR relocations (#965)" (#1249)
This reverts commit fd5e1a482bb25403b59d9176a5c1b9fa1393bcd6.

I'm in the process of stabilizing the 6.12 kernel release and at first,
I thought it would make sense for this to go in at the same time, to
offset some of the size increase from 6.12. However, some new test
failures that need to be debugged have popped up, so I think it makes
sense to get 6.12 stable first.

Co-authored-by: Ben Hillis <benhill@ntdev.microsoft.com>