TDX: LDTR and TR access emulation (#1237)
This fills in the todo for emulating these 4 instructions, as well as
including some small refactors to the previous 4, and adding a missing
security check for load instructions. I don't have a way to test this
end-to-end currently, but the refactors at least don't break the code
that I can test.