flowey: simplify some variable patterns (#1243)
* Add a method for turning a `WriteVar<SideEffect>` into a
`WriteVar<T>`, for throwing away the results of an operation but
retaining the side effect.

* Allow collections (tuples, vecs, maps) of `ClaimedReadVar` to be read
as one, like we allow collections of `ReadVar` to be claimed as one.

* Make `SideEffect` an alias for `()` rather than an uninhabited type so
that it can be read and returned from Rust step functions. This is not
yet used for anything useful, but future cleanup will leverage this.