vmbus_server: release proxy channels once removed by the server. (#1296)
If the driver revokes a channel first, and then the channel gets removed
by `ServerTask` when it's released by the guest, the `ProxyTask` would
never release the channel with the driver, potentially leaving a bunch
of revoked channels in the driver's channel list.

This change updates the channel removal handling to release the channel
if the driver has already revoked it.

I also changed a warning trace to info because it's a noisy warning that
doesn't normally indicate a problem.