virt_mshv_vtl, snp: Add more automatic exits to boot the debug build (#1324)
On the hardware I have that debug assert fires for the automatic exit
that's not in the match statement.

Update the match list to be able to boot the debug build on the newer
hardware.

While at it, convert the exit codes to hex like in the AMD PPR, add
additional codes for easier debugging,
and panic on the negative exit codes that are not covered by the match
statement.