flowey: support directories in typed artifacts (#1277)
Most artifacts contain an explicit list of files and are produced and
consumed as such. But some artifacts, such as the output of rustdoc, are
produced and consumed as an opaque directory. Today, only untyped
artifacts support this.

Support this in typed artifacts by also storing a marker file indicating
that this is an opaque directory.