inspect: clean up field formatting (#1331)
Make a field's formatting (currently hex, binary, and counter) easier to
define. Instead of requiring the formatting to be specified on exactly
the value to be formatted, specify it at any part of the node hierarchy,
and inherit it on all child fields.

This and some other cleanups makes it possible to do a few things:

* You can specify `hex` on the derive of an entire struct in order to
avoid having to specify it on each field.

* You can specify `hex` on fields that already have a `with` or
`iter_by_index`, to avoid needing to manually apply `AsHex` in a `.map`
call.

* You can specify `hex` on arbitrarily complicated numeric fields such
as `Arc<RwLock<AtomicU32>>`, without needing explicit `Inspect` impls
for `AsHex<T>`.

As part of this, simplify a bunch of inspect derives.