igvm: add sidecar to all non-CVM x64 recipes (#1287)
This change updates the main kernel version and adds sidecar to all x64
non-CVM recipes. The only difference between this version of the kernel
and the previous version, the new version contains sidecar patches which
were previously only in the dev kernel. With this update, there is
currently no delta between dev and main kernel versions.


https://github.com/microsoft/OHCL-Linux-Kernel/releases/tag/rolling-lts%2Fhcl-main%2F6.12.9.3

---------

Co-authored-by: Ben Hillis <benhill@ntdev.microsoft.com>