vmbusproxy: Support saved state (#1361)
This PR integrates the `vmbus_server` saved state support with
`proxyintegration`. Previously, `proxyintegration` and vmbusproxy had no
concept of save and restore, and their respective states were completely
reconstructed from scratch on each start.

On start, if using the proxy, the server sends its saved state to the
proxy integration, which in turn repopulates the saved channels and
GPADLs in vmbusproxy. On start, the saved state is cleared.

- Create a new mesh channel in `vmbus_server` and `proxyintegration`
which allows the server to send and clear saved state to the proxy.
- On receive of the saved state, create skeleton channels in vmbusproxy
with their GPADLs to be later claimed by a device.
- On receiving an offer from vmbusproxy, send
`ChannelServerRequest::Restore to the server to restore that channel
with the interrupt received from vmbusproxy.
- On start, tell vmbusproxy to revoke all the unclaimed channels it has
in its list.
- Introduce various new IOCTLs to accomplish the above.
- Introduce new functions in vmbus `SavedState` to allow the proxy to
inspect the `SavedState`.

This PR also fixes a panic related to the Cancel changes in
proxyintegration.