Repo: Update to Rust 1.87 (#1372)
Also cleans up the last few lints needed to be clean on 1.87 that
weren't backwards-compatible with 1.86. Mostly this is adding expects
for clippy::large_enum_variant, which is now able to process more types.
I stared at all the cases for a few minutes, and I don't think they're
in places that are impactful enough to be worth changing.