Repo: Clean up some lint attributes (#1456)
Notably this removes the crate-level expect(unsafe_code) from vmcore,
replacing it with a much more narrowly scoped expect, and adding
forbid(unsafe_code) to all its other modules. This expect is needed due
to our use of linkme inside the implementation of deriving
SavedStateRoot. It is not needed on our derivations of this trait
elsewhere, as the lint is suppressed when it comes from a macro
invocation from a different crate. This is the only place we derive
SavedStateRoot in vmcore itself.