CI: update mirror repo workflow to update submodule in the 2505 staging branch (#1466)
Now that we are in ask mode for the upcoming release, this change
updates the mirroring workflow to submit OSS changes to a staging branch
instead of directly to the release branch. This means that we will need
to periodically merge staging into release (after getting approval).
This is the same flow we did for the 2411 release.

Co-authored-by: Ben Hillis <benhill@ntdev.microsoft.com>