underhill_core: serialize sidecar VP online (#1443) (#1470)
The Linux kernel serializes CPU hotplug. If multiple sidecar VPs need to
be onlined into OpenVMM simultaneously, they will all stop running the
guest while associated Linux threads call into the Linux kernel to
online the CPU (which will block on the CPU hotplug lock or whatever).

This means the average blackout time for a VP that's onlined early in
boot is linear in the number of early-onlined VPs. And thanks to typical
device configurations, this is usually linear in the total number of
VPs. This is a performance problem.

To avoid this, explicitly serialize VP online _before_ the target VP is
stopped. This allows the VP to continue running the guest until it
reaches the front of the online queue. This reduces the average blackout
time to just the time to online one CPU, meaning this solution should
scale to any number of VPs.

Cherry-pick of #1443

Co-authored-by: John Starks <jostarks@microsoft.com>