virt_support_apic: Remove vector tracing filter (#1469)
This removes functionality added in
https://github.com/microsoft/openvmm/commit/7278a20daf243507baab1a912b1bd335a8c534d0
to avoid hitting https://github.com/tokio-rs/tracing/issues/2519. While
the functionality is nice to have, it is not so important as to be worth
potentially dropping events, and there is no performant way to implement
it that can avoid this bug.

Then ban the tracing::enabled macro codebase-wide.