firmware_uefi: Backport EFI Diagnostics (#1498)
(This is a backport PR)

This PR adds EFI Diagnostics, which is a service used to parse UEFI
diagnostics data from an in-memory buffer and send it to our tracing
facilities.

The UEFI firmware will write the GPA of the advanced logger buffer to an
Io port intercept called `SET_EFI_DIAGNOSTICS_GPA`.

The diagnostics service is responsible for reading guest memory at the
specified GPA and parsing the data. This gets triggered when the UEFI
firmware writes to an Io port intercept called
`PROCESS_EFI_DIAGNOSTICS`.

The `PROCESS_EFI_DIAGNOSTICS` UefiCommand gets triggered by the
following conditions:
- UEFI encounters a failure (guest driven via `PROCESS_EFI_DIAGNOSTICS`)
- UEFI fails to boot any device (guest driven via
`PROCESS_EFI_DIAGNOSTICS`)
- UEFI reaches exit boot services

The simplest way to test this is to run:
```
cargo run -- --uefi
```