1.6 cvm guest vsm: track vtl 1 permissions in bitmaps (#1457) (#1487)
When VTL 2 accesses VTL 0 memory on behalf of VTL 0, it needs to be able
to check whether VTL 1 has restricted access to the memory. This change
introduces tracking of VTL 1 permissions using bitmaps and adds some of
the enforcement of these permissions.

Tested:
SNP +/- guest VSM boots
TVM and TDX VMs also boot