tdx: init APs with the ACPI mailbox protocol (#970)
## Mailbox
This is the implementation of #610, initialization of APs with an
implementation of the ACPI Mailbox protocol, described in section
5.2.12.19 of the [ACPI
spec](https://uefi.org/sites/default/files/resources/ACPI_Spec_6_5_Aug29.pdf).
This moves the AP initialization logic for TDX into the shim.

## Changes
**Igvmfilegen**
The TDX reset vector has been modified to spin on APs until the kernel
writes a wakeup command to the mailbox page. The AP then jumps to the
wakeup address written by the kernel.

**Boot Shim**
The kernel expects that APs are ready to be woken up via the wakeup
command, meaning that the VPs must be spinning before we jump to the
kernel. Starting VPs under a HyperV host VMM is done with hypercalls,
but there is no existing method to invoke hypercalls from the shim when
TDX-isolated.

---------

Co-authored-by: Chris Oo <coo@umich.edu>